+++ tests/scheduler/backends/test_qblox_backend.py
@@ -877,7 +877,7 @@ def test_staircase_qasm_prog(start_amp, final_amp):
         init_amp = init_amp - constants.REGISTER_SIZE
 
     final_amp_imm = amp_step_used * (steps_taken - 1) + init_amp
-    awg_output_volt = qcm.awg_output_volt
+    awg_output_volt = qcm.static_hw_properties.max_awg_output_voltage
 
     final_amp_volt = 2 * final_amp_imm / constants.IMMEDIATE_SZ_OFFSET * awg_output_volt
     assert final_amp_volt == pytest.approx(final_amp, 1e-3)
@@ -1041,7 +1041,7 @@ def test_verify_qblox_instruments_version():
     assert (
         wrong_version.value.args[0]
         == f"Qblox DriverVersionError: Installed driver version {nonsense_version} "
-        f"not supported by backend. Please install version 0.4.0 to continue to use "
+        f"not supported by backend. Please install version 0.5.0 to continue to use "
         f"this backend."
     )
 
@@ -1169,8 +1169,14 @@ def test_markers():
             on_marker = int(re.findall(r"\d+", matches[0])[0])
             off_marker = int(re.findall(r"\d+", matches[1])[0])
 
-            assert on_marker == device_compiler.marker_configuration["start"]
-            assert off_marker == device_compiler.marker_configuration["end"]
+            assert (
+                on_marker
+                == device_compiler.static_hw_properties.marker_configuration.start
+            )
+            assert (
+                off_marker
+                == device_compiler.static_hw_properties.marker_configuration.end
+            )
 
     _confirm_correct_markers(program["qcm0"], QcmModule)
     _confirm_correct_markers(program["qrm0"], QrmModule)