+++ CHANGELOG.rst
@@ -11,7 +11,6 @@ Unreleased
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
-* Mixer corrections in Qblox backend are broken because of the switch to real-time modulation. The support of mixer corrections in the Qblox firmware is under development.
 * The `schedules.timedomain_schedules.allxy_sched` function no longer accepts the string "All" as an argument to the `element_select_idx` keyword.
 * The `QuantumDevice.cfg_nr_averages` parameter was renamed to `QuantumDevice.cfg_sched_repetitions`
 * The call signature of `gettables.ScheduleVectorAcqGettable` has been renamed to `gettables.ScheduleGettableSingleChannel`, and the call signature has been updated according to #36 to no longer accept several keyword arguments.
@@ -28,12 +27,14 @@ Merged branches and closed issues
 * InstrumentCoordinator - CompiledSchedule class added to specify interfaces of InstrumentCoordinator and compilation functions (#174, !177).
 * InstrumentCoordinator - CompiledSchedule.last_schedule method added to provide access to last executed schedule (#167, !177).
 * Qblox Backend - Added support for qblox_instruments version 0.4.0 (new acquisition path) (!143).
+* Qblox Backend - Added support for real time mixer corrections rather than pre-distorting the uploaded waveforms (!192).
 * Qblox Backend - Waveforms are now compared using the normalized data array rather than the parameterized description (!182).
 * Qblox Backend - Support for append bin mode (#184, !180).
 * Qblox Backend - Support for using real value pulses on arbitrary outputs added (!142).
 * Updated existing schedules to make use of the acquisition index (#180, !180).
 * Added a function to extract acquisition metadata from a schedule (#179, !180).
 * Qblox ICCs - Compensated integration time for Qblox QRM IC component (!199).
+* Qblox ICCs - Added error handling for error flags given by `get_sequencer_state` (!215)
 * Visualization - Allow user defined axis for plotting circuit diagram (!206)
 
 0.4.0 InstrumentCoordinator and improvements to backends (2021-08-06)